Updated to np.float64 in parameter_plots.py

+++ ogcore/parameter_plots.py
@@ -301,8 +301,9 @@ def plot_fert_rates(
     fert_fine_pred = fert_func(age_fine_pred)
     age_fine = np.hstack((min_yr, age_fine_pred, max_yr))
     fert_fine = np.hstack((0, fert_fine_pred, 0))
-    age_mid_new = np.linspace(np.float(max_yr) / totpers, max_yr, totpers) - (
-        0.5 * np.float(max_yr) / totpers
+    age_mid_new = (
+        np.linspace(np.float64(max_yr) / totpers, max_yr, totpers) -
+        (0.5 * np.float64(max_yr) / totpers)
     )
 
     fig, ax = plt.subplots()
@@ -369,8 +370,9 @@ def plot_mort_rates_data(
         fig (Matplotlib plot object): plot of mortality rates
 
     """
-    age_mid_new = np.linspace(np.float(max_yr) / totpers, max_yr, totpers) - (
-        0.5 * np.float(max_yr) / totpers
+    age_mid_new = (
+        np.linspace(np.float64(max_yr) / totpers, max_yr, totpers) -
+        (0.5 * np.float64(max_yr) / totpers)
     )
     fig, ax = plt.subplots()
     plt.scatter(
